Renaming csv issue test classes and minor cleanup
