+++ tests/MiniExcel.Csv.Tests/Issues/GithubIssuesTests.cs
@@ -1,125 +1,12 @@
 ﻿namespace MiniExcelLib.Csv.Tests.Issues;
 
-public class IssueTests
+public class GithubIssuesTests
 {
     private readonly CsvExporter _csvExporter = MiniExcel.Exporters.GetCsvExporter();
     private readonly CsvImporter _csvImporter = MiniExcel.Importers.GetCsvImporter();
 
     private readonly OpenXmlExporter _openXmlExporter = MiniExcel.Exporters.GetOpenXmlExporter();
     private readonly OpenXmlImporter _openXmlImporter = MiniExcel.Importers.GetOpenXmlImporter();
-
-    [Fact]
-    public void TestPullRequest10()
-    {
-        var path = PathHelper.GetFile("csv/TestIssue142.csv");
-        var config = new CsvConfiguration
-        {
-            SplitFn = row => Regex.Split(row, "[\t,](?=(?:[^\"]|\"[^\"]*\")*$)")
-                .Select(s => Regex.Replace(s.Replace("\"\"", "\""), "^\"|\"$", ""))
-                .ToArray()
-        };
-        
-        var rows = _csvImporter.Query(path, configuration: config).ToList();
-        Assert.Equal(2, rows.Count);
-        Assert.Equal(15, ((IDictionary<string, object>)rows[0]).Count);
-    }
-
-    // https://gitee.com/dotnetchina/MiniExcel/issues/I4X92G
-    [Fact]
-    public void TestIssueI4X92G()
-    {
-        using var file = AutoDeletingPath.Create(ExcelType.Csv);
-        var path = file.ToString();
-
-        {
-            var value = new[]
-            {
-                new { ID = 1, Name = "Jack", InDate = new DateTime(2021,01,03)},
-                new { ID = 2, Name = "Henry", InDate = new DateTime(2020,05,03)}
-            };
-            _csvExporter.Export(path, value);
-            var content = File.ReadAllText(path);
-            Assert.Equal(
-                """
-                 ID,Name,InDate
-                 1,Jack,"2021-01-03 00:00:00"
-                 2,Henry,"2020-05-03 00:00:00"
-
-                 """,
-                content);
-        }
-        {
-            var value = new { ID = 3, Name = "Mike", InDate = new DateTime(2021, 04, 23) };
-            var rowsWritten = _csvExporter.Append(path, value);
-            Assert.Equal(1, rowsWritten);
-
-            var content = File.ReadAllText(path);
-            Assert.Equal(
-                """
-                 ID,Name,InDate
-                 1,Jack,"2021-01-03 00:00:00"
-                 2,Henry,"2020-05-03 00:00:00"
-                 3,Mike,"2021-04-23 00:00:00"
-
-                 """,
-                content);
-        }
-        {
-            var value = new[]
-            {
-                new { ID=4,Name ="Frank",InDate=new DateTime(2021,06,07)},
-                new { ID=5,Name ="Gloria",InDate=new DateTime(2022,05,03)},
-            };
-            var rowsWritten = _csvExporter.Append(path, value);
-            Assert.Equal(2, rowsWritten);
-
-            var content = File.ReadAllText(path);
-            Assert.Equal(
-                """
-                 ID,Name,InDate
-                 1,Jack,"2021-01-03 00:00:00"
-                 2,Henry,"2020-05-03 00:00:00"
-                 3,Mike,"2021-04-23 00:00:00"
-                 4,Frank,"2021-06-07 00:00:00"
-                 5,Gloria,"2022-05-03 00:00:00"
-
-                 """,
-                content);
-        }
-    }
-
-    [Fact]
-    public void TestIssueI4Wda9()
-    {
-        using var path = AutoDeletingPath.Create(ExcelType.Csv);
-        var value = new DataTable();
-        {
-            value.Columns.Add("\"name\"");
-            value.Rows.Add("\"Jack\"");
-        }
-
-        _csvExporter.Export(path.ToString(), value);
-        Assert.Equal("\"\"\"name\"\"\"\r\n\"\"\"Jack\"\"\"\r\n", File.ReadAllText(path.ToString()));
-    }
-
-    // Using stream.SaveAs will close the Stream automatically when Specifying excelType
-    // https://gitee.com/dotnetchina/MiniExcel/issues/I57WMM
-    [Fact]
-    public void TestIssueGiteeI57()
-    {
-        Dictionary<string, object>[] sheets = [new() { ["ID"] = "0001", ["Name"] = "Jack" }];
-        using var stream = new MemoryStream();
-
-        var config = new CsvConfiguration { StreamWriterFunc = x => new StreamWriter(x, Encoding.Default, leaveOpen: true) };
-        _csvExporter.Export(stream, sheets, configuration: config);
-        stream.Seek(0, SeekOrigin.Begin);
-
-        // convert stream to string
-        using var reader = new StreamReader(stream);
-        var text = reader.ReadToEnd();
-
-        Assert.Equal("ID,Name\r\n0001,Jack\r\n", text);
-    }
 
     // Support for Enum Mapping
     [Fact]
@@ -523,7 +410,7 @@ public class IssueTests
     {
         var path = PathHelper.GetFile("/csv/TestIssue298.csv");
 #pragma warning disable CS0618 // Type or member is obsolete
-        var dt = _csvImporter.QueryAsDataTable(path);
+        using var dt = _csvImporter.QueryAsDataTable(path);
 #pragma warning restore CS0618
         Assert.Equal(["ID", "Name", "Age"], dt.Columns.Cast<DataColumn>().Select(x => x.ColumnName));
     }
@@ -676,7 +563,7 @@ public class IssueTests
         //Problem with multi-line when using Query func
         //https://github.com/mini-software/MiniExcel/issues/507
 
-        var path = Path.Combine(Path.GetTempPath(), string.Concat(nameof(IssueTests), "_", nameof(Issue507_1), ".csv"));
+        var path = Path.Combine(Path.GetTempPath(), string.Concat(nameof(GithubIssuesTests), "_", nameof(Issue507_1), ".csv"));
         var values = new Issue507V01[]
         {
             new() { A = "Github", B = DateTime.Parse("2021-01-01"), C = "abcd", D = 123 },
@@ -717,11 +604,10 @@ public class IssueTests
         File.Delete(path);
     }
 
-    //https://github.com/mini-software/MiniExcel/issues/507
     [Fact]
     public void Issue507_2()
     {
-        var path = Path.Combine(Path.GetTempPath(), string.Concat(nameof(IssueTests), "_", nameof(Issue507_2), ".csv"));
+        var path = Path.Combine(Path.GetTempPath(), string.Concat(nameof(GithubIssuesTests), "_", nameof(Issue507_2), ".csv"));
         var values = new Issue507V02[]
         {
             new() { B = DateTime.Parse("2021-01-01"), D = 123 },
